Move tests to new location.
